#40 Performance optimization: Matrix4x4 all to inline
