+++ Libs/ecm/ecm.vcxproj.filters
@@ -18,7 +18,7 @@
     <ClCompile Include="source\math\vector2.cpp" />
     <ClCompile Include="source\math\vector3.cpp" />
     <ClCompile Include="source\math\vector4.cpp" />
-    <ClCompile Include="source\math\matrix4x4.cpp" />
+    <ClCompile Include="ecm\math\matrix4x4.inl" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ecm\ecm_api.h" />
@@ -37,7 +37,7 @@
     <ClInclude Include="ecm\ecm_math.hpp" />
     <ClInclude Include="ecm\math\vector3.h" />
     <ClInclude Include="ecm\math\vector4.h" />
-    <ClInclude Include="ecm\math\matrix4x4.h" />
+    <ClInclude Include="ecm\math\matrix4x4.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="ecm.rc" />
